manless/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
sqlalchemy==1.4.53
pymysql==1.1.0
cryptography==41.0.7
python-multipart==0.0.6
opencv-python==4.8.1.78
pyserial==3.5
pillow==10.1.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
databases[mysql]==0.8.0
alembic==1.13.1
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
httpx==0.25.2
aiohttp==3.9.1
schedule==1.2.0
psutil==5.9.6 